Protecting Video Routes

+++ controllers/videoController.js
@@ -64,7 +64,11 @@ export const getEditVideo = async (req, res) => {
     } = req;
     try {
       const video = await Video.findById(id);
-      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+      if (video.creator != req.user.id) { // router protect
+        throw Error();
+      } else {
+        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+      }
     } catch (error) {
       res.redirect(routes.home);
     }
@@ -88,9 +92,14 @@ export const getEditVideo = async (req, res) => {
         params: { id }
     } = req;
     try {
+      const video = await Video.findById(id);
+      if (video.creator != req.user.id) {
+        throw Error();
+      } else {
         await Video.findOneAndRemove({ _id: id });
+      }
     } catch (error) {
-      //
+      console.log(error);
     }
     res.redirect(routes.home);
   };
